Avoid copying whole directory on Open in all cases
Alternative to #7904, tries to distinguish between

* click on a single ino file downloaded from the internet (old behaviour applies, directory is created an the file is moved in it)
* click on an ino file that for some reason is part of a project with the wrong name (eg. it was downloaded as zip from github, so the containing folder names becomes "projectName-master"). In this case if the directory is created one level above the selected file, all files are moved there and the original directory is recursively deleted)

+++ app/src/processing/app/Editor.java
@@ -1807,7 +1807,14 @@ public class Editor extends JFrame implements RunnerListener {
         }
 
         // create properly named folder
-        File properFolder = new File(sketchFile.getParent(), properParent);
+        File properFolder;
+        if (onlyContainsSketchFiles(sketchFile.getParent())) {
+          // Move the whole folder, so properFolder needs to be created one level above
+          properFolder = new File(new File(sketchFile.getParent()).getParent(), properParent);
+        } else {
+          properFolder = new File(sketchFile.getParent(), properParent);
+        }
+
         if (properFolder.exists()) {
           Base.showWarning(tr("Error"), I18n.format(tr("A folder named \"{0}\" already exists. " +
             "Can't open sketch."), properParent), null);
@@ -1821,14 +1828,20 @@ public class Editor extends JFrame implements RunnerListener {
         // copy the sketch inside
         File properPdeFile = new File(properFolder, sketchFile.getName());
         try {
-          FileUtils.copy(new File(sketchFile.getParent()), properFolder);
+          if (onlyContainsSketchFiles(sketchFile.getParent())) {
+            File dir = new File(sketchFile.getParent());
+            FileUtils.copy(dir, properFolder);
+            // remove the original folder, so user doesn't get confused
+            FileUtils.recursiveDelete(dir);
+          } else {
+            Base.copyFile(sketchFile, properPdeFile);
+            // remove the original file, so user doesn't get confused
+            sketchFile.delete();
+          }
         } catch (IOException e) {
           Base.showWarning(tr("Error"), tr("Could not copy to a proper location."), e);
           return false;
         }
-
-        // remove the original file, so user doesn't get confused
-        sketchFile.delete();
 
         // update with the new path
         file = properPdeFile;
@@ -1868,6 +1881,40 @@ public class Editor extends JFrame implements RunnerListener {
     };
 
     // opening was successful
+    return true;
+  }
+
+  private boolean allowedExtension(String fileName, String[] validExtensions) {
+    int i = fileName.lastIndexOf('.');
+    if (i > 0) {
+        String ext = fileName.substring(i+1).toLowerCase();
+        if (!Arrays.asList(validExtensions).contains(ext)) {
+          return false;
+        }
+    }
+    // no extension or valid extension
+    return true;
+  }
+
+  private boolean onlyContainsSketchFiles(String path) {
+    File folder = new File(path);
+    File[] listOfFiles = folder.listFiles();
+
+    for (int i = 0; i < listOfFiles.length; i++) {
+      String name = listOfFiles[i].getName();
+      if (listOfFiles[i].isFile()) {
+        // allowed files extensions are only .ino, .h, .hpp, .c, .cpp (all cases)
+        String[] valid = {"ino", "h", "hpp" , "c" , "cpp", "pde"};
+        if (!allowedExtension(name, valid)) {
+          return false;
+        }
+      } else if (listOfFiles[i].isDirectory()) {
+        // allowed dir names are only src and extras , plus source control folders
+        if (name != "src" && name != "extras" && !FileUtils.SOURCE_CONTROL_FOLDERS.contains(name)) {
+          return false;
+        }
+      }
+    }
     return true;
   }
